requirements.txt
# Web framework
flask==2.2.3
gunicorn==20.1.0
werkzeug==2.2.3

# Data processing and machine learning
tensorflow==2.12.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2

# Image processing
pillow==9.4.0
opencv-python-headless==4.7.0.72
matplotlib==3.7.1

# API utilities
requests==2.28.2